requirements.txt
onnx~=1.12.0
onnx-simplifier~=0.4.8
